+++ kntest/cmd/kntest/main.go
@@ -17,9 +17,9 @@ limitations under the License.
 package main
 
 import (
-	"log"
+    "log"
 
-	"github.com/spf13/cobra"
+    "github.com/spf13/cobra"
 
 	"knative.dev/test-infra/kntest/pkg/cluster"
 )
