+++ kntest/pkg/cluster/gke/ops/get.go
